Trade_Base_Extendido/requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
ccxt==4.2.24
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
constantly==23.10.4
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.47.2
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.1
incremental==22.10.0
ipykernel==6.29.0
ipython==8.20.0
jaraco.classes==3.3.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keyring==24.3.0
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
more-itertools==10.2.0
multidict==6.0.4
nbformat==5.9.2
nest_asyncio==1.6.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.2
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.42
psutil==5.9.8
pure-eval==0.2.2
pycares==4.4.0
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-okx==0.2.6
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
tenacity==8.2.3
tornado==6.3.3
traitlets==5.14.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
wheel==0.42.0
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1
et-xmlfile==1.1.0
openpyxl==3.1.2
